+++ app/src/main/java/org/simple/clinic/scheduleappointment/facilityselection/FacilitySelectionActivity.kt
@@ -7,22 +7,22 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import io.github.inflationx.viewpump.ViewPumpContextWrapper
 import io.reactivex.Observable
-import io.reactivex.subjects.PublishSubject
+import io.reactivex.rxkotlin.ofType
 import kotlinx.android.synthetic.main.activity_select_facility.*
 import org.simple.clinic.ClinicApp
 import org.simple.clinic.R
-import org.simple.clinic.bindUiToController
+import org.simple.clinic.ReportAnalyticsEvents
 import org.simple.clinic.di.InjectorProviderContextWrapper
 import org.simple.clinic.facility.Facility
+import org.simple.clinic.mobius.MobiusDelegate
 import org.simple.clinic.util.LocaleOverrideContextWrapper
+import org.simple.clinic.util.unsafeLazy
 import org.simple.clinic.util.wrap
-import org.simple.clinic.widgets.ScreenCreated
-import org.simple.clinic.widgets.ScreenDestroyed
 import org.simple.clinic.widgets.UiEvent
 import java.util.Locale
 import javax.inject.Inject
 
-class FacilitySelectionActivity : AppCompatActivity() {
+class FacilitySelectionActivity : AppCompatActivity(), FacilitySelectionUi, FacilitySelectionUiActions {
 
   companion object {
     const val EXTRA_SELECTED_FACILITY = "selected_facility"
@@ -36,9 +36,25 @@ class FacilitySelectionActivity : AppCompatActivity() {
   lateinit var locale: Locale
 
   @Inject
-  lateinit var controller: FacilitySelectionActivityController
+  lateinit var effectHandlerFactory: FacilitySelectionEffectHandler.Factory
 
-  private val onDestroys = PublishSubject.create<ScreenDestroyed>()
+  private val events by unsafeLazy {
+    facilityClicks()
+        .compose(ReportAnalyticsEvents())
+  }
+
+  private val delegate by unsafeLazy {
+    val uiRenderer = FacilitySelectionUiRenderer(this)
+
+    MobiusDelegate.forActivity(
+        events = events.ofType(),
+        defaultModel = FacilitySelectionModel(),
+        update = FacilitySelectionUpdate(),
+        effectHandler = effectHandlerFactory.create(this).build(),
+        init = FacilitySelectionInit(),
+        modelUpdateListener = uiRenderer::render
+    )
+  }
 
   private lateinit var component: FacilitySelectionActivityComponent
 
@@ -46,17 +62,9 @@ class FacilitySelectionActivity : AppCompatActivity() {
     super.onCreate(savedInstanceState)
     setContentView(R.layout.activity_select_facility)
 
-    bindUiToController(
-        ui = this,
-        events = Observable.merge(
-            screenCreates(),
-            facilityClicks()
-        ),
-        controller = controller,
-        screenDestroys = onDestroys
-    )
-
     facilityPickerView.backClicked = this@FacilitySelectionActivity::finish
+
+    delegate.onRestoreInstanceState(savedInstanceState)
   }
 
   override fun attachBaseContext(baseContext: Context) {
@@ -75,12 +83,20 @@ class FacilitySelectionActivity : AppCompatActivity() {
     super.attachBaseContext(wrappedContext)
   }
 
-  override fun onDestroy() {
-    onDestroys.onNext(ScreenDestroyed())
-    super.onDestroy()
+  override fun onStart() {
+    super.onStart()
+    delegate.start()
   }
 
-  private fun screenCreates() = Observable.just<UiEvent>(ScreenCreated())
+  override fun onStop() {
+    delegate.stop()
+    super.onStop()
+  }
+
+  override fun onSaveInstanceState(outState: Bundle) {
+    delegate.onSaveInstanceState(outState)
+    super.onSaveInstanceState(outState)
+  }
 
   private fun facilityClicks(): Observable<UiEvent> {
     return Observable.create { emitter ->
@@ -88,7 +104,7 @@ class FacilitySelectionActivity : AppCompatActivity() {
     }
   }
 
-  fun sendSelectedFacility(selectedFacility: Facility) {
+  override fun sendSelectedFacility(selectedFacility: Facility) {
     val intent = Intent()
     intent.putExtra(EXTRA_SELECTED_FACILITY, selectedFacility)
     setResult(Activity.RESULT_OK, intent)