+++ app/src/main/java/org/simple/clinic/scheduleappointment/facilityselection/FacilitySelectionEffectHandler.kt
@@ -1,0 +1,25 @@
+package org.simple.clinic.scheduleappointment.facilityselection
+
+import com.spotify.mobius.rx2.RxMobius
+import com.squareup.inject.assisted.Assisted
+import com.squareup.inject.assisted.AssistedInject
+import io.reactivex.ObservableTransformer
+import org.simple.clinic.util.scheduler.SchedulersProvider
+
+class FacilitySelectionEffectHandler @AssistedInject constructor(
+    private val schedulers: SchedulersProvider,
+    @Assisted private val uiActions: FacilitySelectionUiActions
+) {
+
+  @AssistedInject.Factory
+  interface Factory {
+    fun create(uiActions: FacilitySelectionUiActions): FacilitySelectionEffectHandler
+  }
+
+  fun build(): ObservableTransformer<FacilitySelectionEffect, FacilitySelectionEvent> {
+    return RxMobius
+        .subtypeEffectHandler<FacilitySelectionEffect, FacilitySelectionEvent>()
+        .addConsumer(ForwardSelectedFacility::class.java, { uiActions.sendSelectedFacility(it.facility) }, schedulers.ui())
+        .build()
+  }
+}