+++ app/src/main/java/org/simple/clinic/scheduleappointment/facilityselection/FacilitySelectionModel.kt
@@ -1,0 +1,7 @@
+package org.simple.clinic.scheduleappointment.facilityselection
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+class FacilitySelectionModel: Parcelable