+++ app/src/main/java/org/simple/clinic/scheduleappointment/facilityselection/FacilitySelectionUi.kt
@@ -1,0 +1,3 @@
+package org.simple.clinic.scheduleappointment.facilityselection
+
+interface FacilitySelectionUi