+++ CHANGELOG.md
@@ -9,9 +9,9 @@
 - Extracted facility selection UI into a separate component
 - Migrated `ForgotPinCreateNewPinScreen` to Mobius
 - Remove `AddToPatientSearch` screens and merge them with `PatientSearch` screens
+- Migrated `FacilitySelectionActivity` to Mobius
 - [In Progress: 30th Jul 2020] Migrate `OverdueScreen` to Mobius
 - [In Progress: 29th Jul 2020] Migrate `ForgotPinConfirmPinScreen` to Mobius
-- [In Progress: 30th Jul 2020] Migrate `FacilitySelectionActivity` to Mobius
 
 ## On Demo
 ### Features